Convert logic moved to seprarate modules and generalized.
