+++ package.json
@@ -24,11 +24,12 @@
     "validate": "npm-run-all build"
   },
   "devDependencies": {
+    "commander": "3.0.0",
+    "copyfiles": "2.1.1",
     "npm-run-all": "4.1.5",
     "prettier": "1.18.2",
     "rimraf": "3.0.0",
-    "typescript": "3.6.2",
-    "commander": "3.0.0",
-    "copyfiles": "2.1.1"
-  }
+    "typescript": "3.6.2"
+  },
+  "dependencies": {}
 }
