+++ src/converters/index.js
@@ -1,0 +1,11 @@
+module.exports.ts = {
+  converters: [require('./typescript-raw'), require('./typescript-css-in-js')],
+  name: 'TypeScript',
+  outFileName: 'index',
+};
+
+module.exports.scss = {
+  converters: [require('./scss')],
+  name: 'SCSS',
+  outFileName: 'tokens',
+};
