+++ src/converters/typescript-css-in-js/interfaces.ts.template
@@ -1,0 +1,5 @@
+export interface DesignTokens {
+  colors: ColorDesignTokens;
+  typography: TypographyDesignTokens;
+  spacing: SpacingDesignTokens;
+}
