+++ src/converters/typescript-raw/interfaces.ts.template
@@ -1,9 +1,3 @@
-export interface DesignTokens {
-  colors: ColorDesignTokens;
-  typography: TypographyDesignTokens;
-  spacing: SpacingDesignTokens;
-}
-
 export interface RawDesignTokens {
   colors: RawColorDesignTokens;
   typography: RawTypographyDesignTokens;
